Provides a list of dates corresponding to each period of the given time series

Provides a list of dates corresponding to each period of the given time series

Usage

daysOf(ts, pos = 1)

Arguments

ts

A time series

pos

The position of the first considered period.

Value

A list of the starting dates of each period

Examples

daysOf(retail$BookStores)
#>   [1] "1992-01-01" "1992-02-01" "1992-03-01" "1992-04-01" "1992-05-01"
#>   [6] "1992-06-01" "1992-07-01" "1992-08-01" "1992-09-01" "1992-10-01"
#>  [11] "1992-11-01" "1992-12-01" "1993-01-01" "1993-02-01" "1993-03-01"
#>  [16] "1993-04-01" "1993-05-01" "1993-06-01" "1993-07-01" "1993-08-01"
#>  [21] "1993-09-01" "1993-10-01" "1993-11-01" "1993-12-01" "1994-01-01"
#>  [26] "1994-02-01" "1994-03-01" "1994-04-01" "1994-05-01" "1994-06-01"
#>  [31] "1994-07-01" "1994-08-01" "1994-09-01" "1994-10-01" "1994-11-01"
#>  [36] "1994-12-01" "1995-01-01" "1995-02-01" "1995-03-01" "1995-04-01"
#>  [41] "1995-05-01" "1995-06-01" "1995-07-01" "1995-08-01" "1995-09-01"
#>  [46] "1995-10-01" "1995-11-01" "1995-12-01" "1996-01-01" "1996-02-01"
#>  [51] "1996-03-01" "1996-04-01" "1996-05-01" "1996-06-01" "1996-07-01"
#>  [56] "1996-08-01" "1996-09-01" "1996-10-01" "1996-11-01" "1996-12-01"
#>  [61] "1997-01-01" "1997-02-01" "1997-03-01" "1997-04-01" "1997-05-01"
#>  [66] "1997-06-01" "1997-07-01" "1997-08-01" "1997-09-01" "1997-10-01"
#>  [71] "1997-11-01" "1997-12-01" "1998-01-01" "1998-02-01" "1998-03-01"
#>  [76] "1998-04-01" "1998-05-01" "1998-06-01" "1998-07-01" "1998-08-01"
#>  [81] "1998-09-01" "1998-10-01" "1998-11-01" "1998-12-01" "1999-01-01"
#>  [86] "1999-02-01" "1999-03-01" "1999-04-01" "1999-05-01" "1999-06-01"
#>  [91] "1999-07-01" "1999-08-01" "1999-09-01" "1999-10-01" "1999-11-01"
#>  [96] "1999-12-01" "2000-01-01" "2000-02-01" "2000-03-01" "2000-04-01"
#> [101] "2000-05-01" "2000-06-01" "2000-07-01" "2000-08-01" "2000-09-01"
#> [106] "2000-10-01" "2000-11-01" "2000-12-01" "2001-01-01" "2001-02-01"
#> [111] "2001-03-01" "2001-04-01" "2001-05-01" "2001-06-01" "2001-07-01"
#> [116] "2001-08-01" "2001-09-01" "2001-10-01" "2001-11-01" "2001-12-01"
#> [121] "2002-01-01" "2002-02-01" "2002-03-01" "2002-04-01" "2002-05-01"
#> [126] "2002-06-01" "2002-07-01" "2002-08-01" "2002-09-01" "2002-10-01"
#> [131] "2002-11-01" "2002-12-01" "2003-01-01" "2003-02-01" "2003-03-01"
#> [136] "2003-04-01" "2003-05-01" "2003-06-01" "2003-07-01" "2003-08-01"
#> [141] "2003-09-01" "2003-10-01" "2003-11-01" "2003-12-01" "2004-01-01"
#> [146] "2004-02-01" "2004-03-01" "2004-04-01" "2004-05-01" "2004-06-01"
#> [151] "2004-07-01" "2004-08-01" "2004-09-01" "2004-10-01" "2004-11-01"
#> [156] "2004-12-01" "2005-01-01" "2005-02-01" "2005-03-01" "2005-04-01"
#> [161] "2005-05-01" "2005-06-01" "2005-07-01" "2005-08-01" "2005-09-01"
#> [166] "2005-10-01" "2005-11-01" "2005-12-01" "2006-01-01" "2006-02-01"
#> [171] "2006-03-01" "2006-04-01" "2006-05-01" "2006-06-01" "2006-07-01"
#> [176] "2006-08-01" "2006-09-01" "2006-10-01" "2006-11-01" "2006-12-01"
#> [181] "2007-01-01" "2007-02-01" "2007-03-01" "2007-04-01" "2007-05-01"
#> [186] "2007-06-01" "2007-07-01" "2007-08-01" "2007-09-01" "2007-10-01"
#> [191] "2007-11-01" "2007-12-01" "2008-01-01" "2008-02-01" "2008-03-01"
#> [196] "2008-04-01" "2008-05-01" "2008-06-01" "2008-07-01" "2008-08-01"
#> [201] "2008-09-01" "2008-10-01" "2008-11-01" "2008-12-01" "2009-01-01"
#> [206] "2009-02-01" "2009-03-01" "2009-04-01" "2009-05-01" "2009-06-01"
#> [211] "2009-07-01" "2009-08-01" "2009-09-01" "2009-10-01" "2009-11-01"
#> [216] "2009-12-01" "2010-01-01" "2010-02-01" "2010-03-01" "2010-04-01"
#> [221] "2010-05-01" "2010-06-01" "2010-07-01" "2010-08-01" "2010-09-01"
#> [226] "2010-10-01" "2010-11-01" "2010-12-01"