rjd3toolkit: Utility Functions Around 'JDemetra+ 3.0'

R Interface to 'JDemetra+ 3.x' (https://github.com/jdemetra) time series analysis software. It provides functions allowing to model time series (create outlier regressors, user-defined calendar regressors, Unobserved Components AutoRegressive Integrated Moving Average (UCARIMA) models...), to test the presence of trading days or seasonal effects and also to set specifications in pre-adjustment and benchmarking when using 'rjd3x13' or 'rjd3tramoseats'.

See also

Author

Maintainer: Tanguy Barthelemy timeserieswithjdemetraandr@gmail.com [artist, copyright holder]

Authors:

Tanguy Barthelemy timeserieswithjdemetraandr@gmail.com [artist, copyright holder]

Jean Palate palatejean@gmail.com

Alain Quartier-la-Tente alain.quartier@yahoo.fr (ORCID)

Anna Smyk anna.smyk@insee.fr