Estimate SARIMA Model

Estimate SARIMA Model

Usage

sarima_estimate(
  x,
  order = c(0, 0, 0),
  seasonal = list(order = c(0, 0, 0), period = NA),
  mean = FALSE,
  xreg = NULL,
  eps = 1e-09
)

Arguments

x

a univariate time series.

order

vector specifying of the non-seasonal part of the ARIMA model: the AR order, the degree of differencing, and the MA order.

seasonal

specification of the seasonal part of the ARIMA model and the seasonal frequency (by default equals to frequency(x)). Either a list with components order and period or a numeric vector specifying the seasonal order (the default period is then used).

mean

should the SARIMA model include an intercept term.

xreg

vector or matrix of external regressors.

eps

precision.

Examples

y <- ABS$X0.2.09.10.M
sarima_estimate(y, order = c(0,1,1), seasonal = c(0,1,1))
#> SARIMA model:  (0,1,1) (0,1,1) [12]
#> 
#> Coefficients
#>           Estimate Std. Error  T-stat
#> theta(1)  -0.87640    0.02290 -38.274
#> btheta(1) -0.38754    0.05071  -7.642
#> 
#> No regression variables