Trading day Regressor for Stock series

Allows to generate a specific regressor for correcting trading days effects in Stock series.

Usage

stock_td(frequency, start, length, s, w = 31)

Arguments

frequency

Frequency of the series, number of periods per year (12, 4, 3, 2...)

start, length

First date (array with the first year and the first period, for instance c(1980, 1)) and number of periods of the output variables. Can also be provided with the s argument

s

time series used to get the dates for the trading days variables. If supplied the parameters frequency, start and length are ignored.

w

indicates day of the month when inventories and other stocks are reported. (to denote the last day of the month enter 31).

Value

Time series (object of class c("ts","mts","matrix")).

Details

The regressor will have the value -1 if the w-th day is a Sunday, 1 if it is a Monday as 0 otherwise.

References

More information on calendar correction in JDemetra+ online documentation: https://jdemetra-new-documentation.netlify.app/a-calendar-correction

See also

Examples

stock_td(frequency = 12L, start = c(1990L, 1L), length = 480L, w = 1L)
#>          Monday Tuesday Wednesday Thursday Friday Saturday
#> Jan 1990      1       0         0        0      0        0
#> Feb 1990      0       0         0        1      0        0
#> Mar 1990      0       0         0        1      0        0
#> Apr 1990     -1      -1        -1       -1     -1       -1
#> May 1990      0       1         0        0      0        0
#> Jun 1990      0       0         0        0      1        0
#> Jul 1990     -1      -1        -1       -1     -1       -1
#> Aug 1990      0       0         1        0      0        0
#> Sep 1990      0       0         0        0      0        1
#> Oct 1990      1       0         0        0      0        0
#> Nov 1990      0       0         0        1      0        0
#> Dec 1990      0       0         0        0      0        1
#> Jan 1991      0       1         0        0      0        0
#> Feb 1991      0       0         0        0      1        0
#> Mar 1991      0       0         0        0      1        0
#> Apr 1991      1       0         0        0      0        0
#> May 1991      0       0         1        0      0        0
#> Jun 1991      0       0         0        0      0        1
#> Jul 1991      1       0         0        0      0        0
#> Aug 1991      0       0         0        1      0        0
#> Sep 1991     -1      -1        -1       -1     -1       -1
#> Oct 1991      0       1         0        0      0        0
#> Nov 1991      0       0         0        0      1        0
#> Dec 1991     -1      -1        -1       -1     -1       -1
#> Jan 1992      0       0         1        0      0        0
#> Feb 1992      0       0         0        0      0        1
#> Mar 1992     -1      -1        -1       -1     -1       -1
#> Apr 1992      0       0         1        0      0        0
#> May 1992      0       0         0        0      1        0
#> Jun 1992      1       0         0        0      0        0
#> Jul 1992      0       0         1        0      0        0
#> Aug 1992      0       0         0        0      0        1
#> Sep 1992      0       1         0        0      0        0
#> Oct 1992      0       0         0        1      0        0
#> Nov 1992     -1      -1        -1       -1     -1       -1
#> Dec 1992      0       1         0        0      0        0
#> Jan 1993      0       0         0        0      1        0
#> Feb 1993      1       0         0        0      0        0
#> Mar 1993      1       0         0        0      0        0
#> Apr 1993      0       0         0        1      0        0
#> May 1993      0       0         0        0      0        1
#> Jun 1993      0       1         0        0      0        0
#> Jul 1993      0       0         0        1      0        0
#> Aug 1993     -1      -1        -1       -1     -1       -1
#> Sep 1993      0       0         1        0      0        0
#> Oct 1993      0       0         0        0      1        0
#> Nov 1993      1       0         0        0      0        0
#> Dec 1993      0       0         1        0      0        0
#> Jan 1994      0       0         0        0      0        1
#> Feb 1994      0       1         0        0      0        0
#> Mar 1994      0       1         0        0      0        0
#> Apr 1994      0       0         0        0      1        0
#> May 1994     -1      -1        -1       -1     -1       -1
#> Jun 1994      0       0         1        0      0        0
#> Jul 1994      0       0         0        0      1        0
#> Aug 1994      1       0         0        0      0        0
#> Sep 1994      0       0         0        1      0        0
#> Oct 1994      0       0         0        0      0        1
#> Nov 1994      0       1         0        0      0        0
#> Dec 1994      0       0         0        1      0        0
#> Jan 1995     -1      -1        -1       -1     -1       -1
#> Feb 1995      0       0         1        0      0        0
#> Mar 1995      0       0         1        0      0        0
#> Apr 1995      0       0         0        0      0        1
#> May 1995      1       0         0        0      0        0
#> Jun 1995      0       0         0        1      0        0
#> Jul 1995      0       0         0        0      0        1
#> Aug 1995      0       1         0        0      0        0
#> Sep 1995      0       0         0        0      1        0
#> Oct 1995     -1      -1        -1       -1     -1       -1
#> Nov 1995      0       0         1        0      0        0
#> Dec 1995      0       0         0        0      1        0
#> Jan 1996      1       0         0        0      0        0
#> Feb 1996      0       0         0        1      0        0
#> Mar 1996      0       0         0        0      1        0
#> Apr 1996      1       0         0        0      0        0
#> May 1996      0       0         1        0      0        0
#> Jun 1996      0       0         0        0      0        1
#> Jul 1996      1       0         0        0      0        0
#> Aug 1996      0       0         0        1      0        0
#> Sep 1996     -1      -1        -1       -1     -1       -1
#> Oct 1996      0       1         0        0      0        0
#> Nov 1996      0       0         0        0      1        0
#> Dec 1996     -1      -1        -1       -1     -1       -1
#> Jan 1997      0       0         1        0      0        0
#> Feb 1997      0       0         0        0      0        1
#> Mar 1997      0       0         0        0      0        1
#> Apr 1997      0       1         0        0      0        0
#> May 1997      0       0         0        1      0        0
#> Jun 1997     -1      -1        -1       -1     -1       -1
#> Jul 1997      0       1         0        0      0        0
#> Aug 1997      0       0         0        0      1        0
#> Sep 1997      1       0         0        0      0        0
#> Oct 1997      0       0         1        0      0        0
#> Nov 1997      0       0         0        0      0        1
#> Dec 1997      1       0         0        0      0        0
#> Jan 1998      0       0         0        1      0        0
#> Feb 1998     -1      -1        -1       -1     -1       -1
#> Mar 1998     -1      -1        -1       -1     -1       -1
#> Apr 1998      0       0         1        0      0        0
#> May 1998      0       0         0        0      1        0
#> Jun 1998      1       0         0        0      0        0
#> Jul 1998      0       0         1        0      0        0
#> Aug 1998      0       0         0        0      0        1
#> Sep 1998      0       1         0        0      0        0
#> Oct 1998      0       0         0        1      0        0
#> Nov 1998     -1      -1        -1       -1     -1       -1
#> Dec 1998      0       1         0        0      0        0
#> Jan 1999      0       0         0        0      1        0
#> Feb 1999      1       0         0        0      0        0
#> Mar 1999      1       0         0        0      0        0
#> Apr 1999      0       0         0        1      0        0
#> May 1999      0       0         0        0      0        1
#> Jun 1999      0       1         0        0      0        0
#> Jul 1999      0       0         0        1      0        0
#> Aug 1999     -1      -1        -1       -1     -1       -1
#> Sep 1999      0       0         1        0      0        0
#> Oct 1999      0       0         0        0      1        0
#> Nov 1999      1       0         0        0      0        0
#> Dec 1999      0       0         1        0      0        0
#> Jan 2000      0       0         0        0      0        1
#> Feb 2000      0       1         0        0      0        0
#> Mar 2000      0       0         1        0      0        0
#> Apr 2000      0       0         0        0      0        1
#> May 2000      1       0         0        0      0        0
#> Jun 2000      0       0         0        1      0        0
#> Jul 2000      0       0         0        0      0        1
#> Aug 2000      0       1         0        0      0        0
#> Sep 2000      0       0         0        0      1        0
#> Oct 2000     -1      -1        -1       -1     -1       -1
#> Nov 2000      0       0         1        0      0        0
#> Dec 2000      0       0         0        0      1        0
#> Jan 2001      1       0         0        0      0        0
#> Feb 2001      0       0         0        1      0        0
#> Mar 2001      0       0         0        1      0        0
#> Apr 2001     -1      -1        -1       -1     -1       -1
#> May 2001      0       1         0        0      0        0
#> Jun 2001      0       0         0        0      1        0
#> Jul 2001     -1      -1        -1       -1     -1       -1
#> Aug 2001      0       0         1        0      0        0
#> Sep 2001      0       0         0        0      0        1
#> Oct 2001      1       0         0        0      0        0
#> Nov 2001      0       0         0        1      0        0
#> Dec 2001      0       0         0        0      0        1
#> Jan 2002      0       1         0        0      0        0
#> Feb 2002      0       0         0        0      1        0
#> Mar 2002      0       0         0        0      1        0
#> Apr 2002      1       0         0        0      0        0
#> May 2002      0       0         1        0      0        0
#> Jun 2002      0       0         0        0      0        1
#> Jul 2002      1       0         0        0      0        0
#> Aug 2002      0       0         0        1      0        0
#> Sep 2002     -1      -1        -1       -1     -1       -1
#> Oct 2002      0       1         0        0      0        0
#> Nov 2002      0       0         0        0      1        0
#> Dec 2002     -1      -1        -1       -1     -1       -1
#> Jan 2003      0       0         1        0      0        0
#> Feb 2003      0       0         0        0      0        1
#> Mar 2003      0       0         0        0      0        1
#> Apr 2003      0       1         0        0      0        0
#> May 2003      0       0         0        1      0        0
#> Jun 2003     -1      -1        -1       -1     -1       -1
#> Jul 2003      0       1         0        0      0        0
#> Aug 2003      0       0         0        0      1        0
#> Sep 2003      1       0         0        0      0        0
#> Oct 2003      0       0         1        0      0        0
#> Nov 2003      0       0         0        0      0        1
#> Dec 2003      1       0         0        0      0        0
#> Jan 2004      0       0         0        1      0        0
#> Feb 2004     -1      -1        -1       -1     -1       -1
#> Mar 2004      1       0         0        0      0        0
#> Apr 2004      0       0         0        1      0        0
#> May 2004      0       0         0        0      0        1
#> Jun 2004      0       1         0        0      0        0
#> Jul 2004      0       0         0        1      0        0
#> Aug 2004     -1      -1        -1       -1     -1       -1
#> Sep 2004      0       0         1        0      0        0
#> Oct 2004      0       0         0        0      1        0
#> Nov 2004      1       0         0        0      0        0
#> Dec 2004      0       0         1        0      0        0
#> Jan 2005      0       0         0        0      0        1
#> Feb 2005      0       1         0        0      0        0
#> Mar 2005      0       1         0        0      0        0
#> Apr 2005      0       0         0        0      1        0
#> May 2005     -1      -1        -1       -1     -1       -1
#> Jun 2005      0       0         1        0      0        0
#> Jul 2005      0       0         0        0      1        0
#> Aug 2005      1       0         0        0      0        0
#> Sep 2005      0       0         0        1      0        0
#> Oct 2005      0       0         0        0      0        1
#> Nov 2005      0       1         0        0      0        0
#> Dec 2005      0       0         0        1      0        0
#> Jan 2006     -1      -1        -1       -1     -1       -1
#> Feb 2006      0       0         1        0      0        0
#> Mar 2006      0       0         1        0      0        0
#> Apr 2006      0       0         0        0      0        1
#> May 2006      1       0         0        0      0        0
#> Jun 2006      0       0         0        1      0        0
#> Jul 2006      0       0         0        0      0        1
#> Aug 2006      0       1         0        0      0        0
#> Sep 2006      0       0         0        0      1        0
#> Oct 2006     -1      -1        -1       -1     -1       -1
#> Nov 2006      0       0         1        0      0        0
#> Dec 2006      0       0         0        0      1        0
#> Jan 2007      1       0         0        0      0        0
#> Feb 2007      0       0         0        1      0        0
#> Mar 2007      0       0         0        1      0        0
#> Apr 2007     -1      -1        -1       -1     -1       -1
#> May 2007      0       1         0        0      0        0
#> Jun 2007      0       0         0        0      1        0
#> Jul 2007     -1      -1        -1       -1     -1       -1
#> Aug 2007      0       0         1        0      0        0
#> Sep 2007      0       0         0        0      0        1
#> Oct 2007      1       0         0        0      0        0
#> Nov 2007      0       0         0        1      0        0
#> Dec 2007      0       0         0        0      0        1
#> Jan 2008      0       1         0        0      0        0
#> Feb 2008      0       0         0        0      1        0
#> Mar 2008      0       0         0        0      0        1
#> Apr 2008      0       1         0        0      0        0
#> May 2008      0       0         0        1      0        0
#> Jun 2008     -1      -1        -1       -1     -1       -1
#> Jul 2008      0       1         0        0      0        0
#> Aug 2008      0       0         0        0      1        0
#> Sep 2008      1       0         0        0      0        0
#> Oct 2008      0       0         1        0      0        0
#> Nov 2008      0       0         0        0      0        1
#> Dec 2008      1       0         0        0      0        0
#> Jan 2009      0       0         0        1      0        0
#> Feb 2009     -1      -1        -1       -1     -1       -1
#> Mar 2009     -1      -1        -1       -1     -1       -1
#> Apr 2009      0       0         1        0      0        0
#> May 2009      0       0         0        0      1        0
#> Jun 2009      1       0         0        0      0        0
#> Jul 2009      0       0         1        0      0        0
#> Aug 2009      0       0         0        0      0        1
#> Sep 2009      0       1         0        0      0        0
#> Oct 2009      0       0         0        1      0        0
#> Nov 2009     -1      -1        -1       -1     -1       -1
#> Dec 2009      0       1         0        0      0        0
#> Jan 2010      0       0         0        0      1        0
#> Feb 2010      1       0         0        0      0        0
#> Mar 2010      1       0         0        0      0        0
#> Apr 2010      0       0         0        1      0        0
#> May 2010      0       0         0        0      0        1
#> Jun 2010      0       1         0        0      0        0
#> Jul 2010      0       0         0        1      0        0
#> Aug 2010     -1      -1        -1       -1     -1       -1
#> Sep 2010      0       0         1        0      0        0
#> Oct 2010      0       0         0        0      1        0
#> Nov 2010      1       0         0        0      0        0
#> Dec 2010      0       0         1        0      0        0
#> Jan 2011      0       0         0        0      0        1
#> Feb 2011      0       1         0        0      0        0
#> Mar 2011      0       1         0        0      0        0
#> Apr 2011      0       0         0        0      1        0
#> May 2011     -1      -1        -1       -1     -1       -1
#> Jun 2011      0       0         1        0      0        0
#> Jul 2011      0       0         0        0      1        0
#> Aug 2011      1       0         0        0      0        0
#> Sep 2011      0       0         0        1      0        0
#> Oct 2011      0       0         0        0      0        1
#> Nov 2011      0       1         0        0      0        0
#> Dec 2011      0       0         0        1      0        0
#> Jan 2012     -1      -1        -1       -1     -1       -1
#> Feb 2012      0       0         1        0      0        0
#> Mar 2012      0       0         0        1      0        0
#> Apr 2012     -1      -1        -1       -1     -1       -1
#> May 2012      0       1         0        0      0        0
#> Jun 2012      0       0         0        0      1        0
#> Jul 2012     -1      -1        -1       -1     -1       -1
#> Aug 2012      0       0         1        0      0        0
#> Sep 2012      0       0         0        0      0        1
#> Oct 2012      1       0         0        0      0        0
#> Nov 2012      0       0         0        1      0        0
#> Dec 2012      0       0         0        0      0        1
#> Jan 2013      0       1         0        0      0        0
#> Feb 2013      0       0         0        0      1        0
#> Mar 2013      0       0         0        0      1        0
#> Apr 2013      1       0         0        0      0        0
#> May 2013      0       0         1        0      0        0
#> Jun 2013      0       0         0        0      0        1
#> Jul 2013      1       0         0        0      0        0
#> Aug 2013      0       0         0        1      0        0
#> Sep 2013     -1      -1        -1       -1     -1       -1
#> Oct 2013      0       1         0        0      0        0
#> Nov 2013      0       0         0        0      1        0
#> Dec 2013     -1      -1        -1       -1     -1       -1
#> Jan 2014      0       0         1        0      0        0
#> Feb 2014      0       0         0        0      0        1
#> Mar 2014      0       0         0        0      0        1
#> Apr 2014      0       1         0        0      0        0
#> May 2014      0       0         0        1      0        0
#> Jun 2014     -1      -1        -1       -1     -1       -1
#> Jul 2014      0       1         0        0      0        0
#> Aug 2014      0       0         0        0      1        0
#> Sep 2014      1       0         0        0      0        0
#> Oct 2014      0       0         1        0      0        0
#> Nov 2014      0       0         0        0      0        1
#> Dec 2014      1       0         0        0      0        0
#> Jan 2015      0       0         0        1      0        0
#> Feb 2015     -1      -1        -1       -1     -1       -1
#> Mar 2015     -1      -1        -1       -1     -1       -1
#> Apr 2015      0       0         1        0      0        0
#> May 2015      0       0         0        0      1        0
#> Jun 2015      1       0         0        0      0        0
#> Jul 2015      0       0         1        0      0        0
#> Aug 2015      0       0         0        0      0        1
#> Sep 2015      0       1         0        0      0        0
#> Oct 2015      0       0         0        1      0        0
#> Nov 2015     -1      -1        -1       -1     -1       -1
#> Dec 2015      0       1         0        0      0        0
#> Jan 2016      0       0         0        0      1        0
#> Feb 2016      1       0         0        0      0        0
#> Mar 2016      0       1         0        0      0        0
#> Apr 2016      0       0         0        0      1        0
#> May 2016     -1      -1        -1       -1     -1       -1
#> Jun 2016      0       0         1        0      0        0
#> Jul 2016      0       0         0        0      1        0
#> Aug 2016      1       0         0        0      0        0
#> Sep 2016      0       0         0        1      0        0
#> Oct 2016      0       0         0        0      0        1
#> Nov 2016      0       1         0        0      0        0
#> Dec 2016      0       0         0        1      0        0
#> Jan 2017     -1      -1        -1       -1     -1       -1
#> Feb 2017      0       0         1        0      0        0
#> Mar 2017      0       0         1        0      0        0
#> Apr 2017      0       0         0        0      0        1
#> May 2017      1       0         0        0      0        0
#> Jun 2017      0       0         0        1      0        0
#> Jul 2017      0       0         0        0      0        1
#> Aug 2017      0       1         0        0      0        0
#> Sep 2017      0       0         0        0      1        0
#> Oct 2017     -1      -1        -1       -1     -1       -1
#> Nov 2017      0       0         1        0      0        0
#> Dec 2017      0       0         0        0      1        0
#> Jan 2018      1       0         0        0      0        0
#> Feb 2018      0       0         0        1      0        0
#> Mar 2018      0       0         0        1      0        0
#> Apr 2018     -1      -1        -1       -1     -1       -1
#> May 2018      0       1         0        0      0        0
#> Jun 2018      0       0         0        0      1        0
#> Jul 2018     -1      -1        -1       -1     -1       -1
#> Aug 2018      0       0         1        0      0        0
#> Sep 2018      0       0         0        0      0        1
#> Oct 2018      1       0         0        0      0        0
#> Nov 2018      0       0         0        1      0        0
#> Dec 2018      0       0         0        0      0        1
#> Jan 2019      0       1         0        0      0        0
#> Feb 2019      0       0         0        0      1        0
#> Mar 2019      0       0         0        0      1        0
#> Apr 2019      1       0         0        0      0        0
#> May 2019      0       0         1        0      0        0
#> Jun 2019      0       0         0        0      0        1
#> Jul 2019      1       0         0        0      0        0
#> Aug 2019      0       0         0        1      0        0
#> Sep 2019     -1      -1        -1       -1     -1       -1
#> Oct 2019      0       1         0        0      0        0
#> Nov 2019      0       0         0        0      1        0
#> Dec 2019     -1      -1        -1       -1     -1       -1
#> Jan 2020      0       0         1        0      0        0
#> Feb 2020      0       0         0        0      0        1
#> Mar 2020     -1      -1        -1       -1     -1       -1
#> Apr 2020      0       0         1        0      0        0
#> May 2020      0       0         0        0      1        0
#> Jun 2020      1       0         0        0      0        0
#> Jul 2020      0       0         1        0      0        0
#> Aug 2020      0       0         0        0      0        1
#> Sep 2020      0       1         0        0      0        0
#> Oct 2020      0       0         0        1      0        0
#> Nov 2020     -1      -1        -1       -1     -1       -1
#> Dec 2020      0       1         0        0      0        0
#> Jan 2021      0       0         0        0      1        0
#> Feb 2021      1       0         0        0      0        0
#> Mar 2021      1       0         0        0      0        0
#> Apr 2021      0       0         0        1      0        0
#> May 2021      0       0         0        0      0        1
#> Jun 2021      0       1         0        0      0        0
#> Jul 2021      0       0         0        1      0        0
#> Aug 2021     -1      -1        -1       -1     -1       -1
#> Sep 2021      0       0         1        0      0        0
#> Oct 2021      0       0         0        0      1        0
#> Nov 2021      1       0         0        0      0        0
#> Dec 2021      0       0         1        0      0        0
#> Jan 2022      0       0         0        0      0        1
#> Feb 2022      0       1         0        0      0        0
#> Mar 2022      0       1         0        0      0        0
#> Apr 2022      0       0         0        0      1        0
#> May 2022     -1      -1        -1       -1     -1       -1
#> Jun 2022      0       0         1        0      0        0
#> Jul 2022      0       0         0        0      1        0
#> Aug 2022      1       0         0        0      0        0
#> Sep 2022      0       0         0        1      0        0
#> Oct 2022      0       0         0        0      0        1
#> Nov 2022      0       1         0        0      0        0
#> Dec 2022      0       0         0        1      0        0
#> Jan 2023     -1      -1        -1       -1     -1       -1
#> Feb 2023      0       0         1        0      0        0
#> Mar 2023      0       0         1        0      0        0
#> Apr 2023      0       0         0        0      0        1
#> May 2023      1       0         0        0      0        0
#> Jun 2023      0       0         0        1      0        0
#> Jul 2023      0       0         0        0      0        1
#> Aug 2023      0       1         0        0      0        0
#> Sep 2023      0       0         0        0      1        0
#> Oct 2023     -1      -1        -1       -1     -1       -1
#> Nov 2023      0       0         1        0      0        0
#> Dec 2023      0       0         0        0      1        0
#> Jan 2024      1       0         0        0      0        0
#> Feb 2024      0       0         0        1      0        0
#> Mar 2024      0       0         0        0      1        0
#> Apr 2024      1       0         0        0      0        0
#> May 2024      0       0         1        0      0        0
#> Jun 2024      0       0         0        0      0        1
#> Jul 2024      1       0         0        0      0        0
#> Aug 2024      0       0         0        1      0        0
#> Sep 2024     -1      -1        -1       -1     -1       -1
#> Oct 2024      0       1         0        0      0        0
#> Nov 2024      0       0         0        0      1        0
#> Dec 2024     -1      -1        -1       -1     -1       -1
#> Jan 2025      0       0         1        0      0        0
#> Feb 2025      0       0         0        0      0        1
#> Mar 2025      0       0         0        0      0        1
#> Apr 2025      0       1         0        0      0        0
#> May 2025      0       0         0        1      0        0
#> Jun 2025     -1      -1        -1       -1     -1       -1
#> Jul 2025      0       1         0        0      0        0
#> Aug 2025      0       0         0        0      1        0
#> Sep 2025      1       0         0        0      0        0
#> Oct 2025      0       0         1        0      0        0
#> Nov 2025      0       0         0        0      0        1
#> Dec 2025      1       0         0        0      0        0
#> Jan 2026      0       0         0        1      0        0
#> Feb 2026     -1      -1        -1       -1     -1       -1
#> Mar 2026     -1      -1        -1       -1     -1       -1
#> Apr 2026      0       0         1        0      0        0
#> May 2026      0       0         0        0      1        0
#> Jun 2026      1       0         0        0      0        0
#> Jul 2026      0       0         1        0      0        0
#> Aug 2026      0       0         0        0      0        1
#> Sep 2026      0       1         0        0      0        0
#> Oct 2026      0       0         0        1      0        0
#> Nov 2026     -1      -1        -1       -1     -1       -1
#> Dec 2026      0       1         0        0      0        0
#> Jan 2027      0       0         0        0      1        0
#> Feb 2027      1       0         0        0      0        0
#> Mar 2027      1       0         0        0      0        0
#> Apr 2027      0       0         0        1      0        0
#> May 2027      0       0         0        0      0        1
#> Jun 2027      0       1         0        0      0        0
#> Jul 2027      0       0         0        1      0        0
#> Aug 2027     -1      -1        -1       -1     -1       -1
#> Sep 2027      0       0         1        0      0        0
#> Oct 2027      0       0         0        0      1        0
#> Nov 2027      1       0         0        0      0        0
#> Dec 2027      0       0         1        0      0        0
#> Jan 2028      0       0         0        0      0        1
#> Feb 2028      0       1         0        0      0        0
#> Mar 2028      0       0         1        0      0        0
#> Apr 2028      0       0         0        0      0        1
#> May 2028      1       0         0        0      0        0
#> Jun 2028      0       0         0        1      0        0
#> Jul 2028      0       0         0        0      0        1
#> Aug 2028      0       1         0        0      0        0
#> Sep 2028      0       0         0        0      1        0
#> Oct 2028     -1      -1        -1       -1     -1       -1
#> Nov 2028      0       0         1        0      0        0
#> Dec 2028      0       0         0        0      1        0
#> Jan 2029      1       0         0        0      0        0
#> Feb 2029      0       0         0        1      0        0
#> Mar 2029      0       0         0        1      0        0
#> Apr 2029     -1      -1        -1       -1     -1       -1
#> May 2029      0       1         0        0      0        0
#> Jun 2029      0       0         0        0      1        0
#> Jul 2029     -1      -1        -1       -1     -1       -1
#> Aug 2029      0       0         1        0      0        0
#> Sep 2029      0       0         0        0      0        1
#> Oct 2029      1       0         0        0      0        0
#> Nov 2029      0       0         0        1      0        0
#> Dec 2029      0       0         0        0      0        1