Set estimation sub-span and quality check specification

Function allowing to check if the series can be processed and to define a sub-span on which estimation will be performed

Usage

set_basic(
  x,
  type = c(NA, "All", "From", "To", "Between", "Last", "First", "Excluding"),
  d0 = NULL,
  d1 = NULL,
  n0 = 0,
  n1 = 0,
  preliminary.check = NA,
  preprocessing = NA
)

Arguments

x

the specification to customize, must be a "SPEC" class object (see details).

type, d0, d1, n0, n1

parameters to specify the sub-span .

d0 and d1 characters in the format "YYYY-MM-DD" to specify first/last date of the span when type equals to "From", "To" or "Between". Date corresponding to d0 will be included in the sub-span Date corresponding to d1 will be excluded from the sub span

n0 and n1 numeric to specify the number of periods at the beginning/end of the series to be used for defining the sub-span (type equals to "First", "Last") or to exclude (type equals to "Excluding").

preliminary.check

a Boolean to check the quality of the input series and exclude highly problematic ones (e.g. the series with a number of identical observations and/or missing values above pre-specified threshold values).

preprocessing

(REGARIMA/X13 Specific) a Boolean to enable/disable the pre-processing. Option disabled for the moment.

Details

x specification param must be a JD3_X13_SPEC" class object generated with rjd3x13::x13_spec() (or "JD3_REGARIMA_SPEC" generated with rjd3x13::spec_regarima() or "JD3_TRAMOSEATS_SPEC" generated with rjd3tramoseats::spec_tramoseats() or "JD3_TRAMO_SPEC" generated with rjd3tramoseats::spec_tramo()).

References

More information in JDemetra+ online documentation: https://jdemetra-new-documentation.netlify.app/

See also

Examples

# init_spec <- rjd3x13::x13_spec("RSA5c")
# estimation on sub-span between two dates (date d1 is excluded)
# new_spec<-set_basic(init_spec,type = "Between",d0 = "2014-01-01",
# d1 = "2019-01-01", preliminary.check = TRUE, preprocessing = TRUE)
# Estimation on the first 60 observations
# new_spec <-set_basic(init_spec,Type="First", n0 = 60,
#                      preliminary.check = TRUE,
#                      preprocessing= TRUE)
# Estimation on the last 60 observations
# new_spec <-set_basic(init_spec,Type="Last", n1 = 60,
#                      preliminary.check = TRUE,
#                      preprocessing= TRUE)
# Estimation excluding 60 observations at the beginning and 36 at the end of the series
# new_spec <-set_basic(init_spec,Type="Excluding", n0=60, n1=36,
#                      preliminary.check = TRUE,
#                      preprocessing= TRUE)